+++ src/Variable.ts
@@ -1,4 +1,10 @@
-export class Variable {
+import { IStatementGiver } from './IStatementGiver'
+
+export class Variable implements IStatementGiver{
 	public constructor(public readonly name: string) {
+	}
+
+	public getStmt(): string {
+		return this.name
 	}
 }